+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## [1.2.0] - 2018-10-10
+
+### Added
+
+- New `parseString` method to parse a string directly
+
+### Changed
+
+- Improved readme and added changelog section
+
 ## [1.1.0] - 2018-08-10
 
 ### Added